Checked for file permissions

+++ sinatra-test-app/vote_logger.rb
@@ -8,12 +8,16 @@ class VoteLogger
   end
 
   def self.vote(number="1", vote="si")
-    temp_filename = Time.now.to_i.to_s
-    File.open(temp_filename, "w") do |f|
-      f << ( @@vote_results << {:number => number, :vote => vote, :created_at => Time.now.to_i} ).to_yaml
+    unless @@dont_write
+      temp_filename = Time.now.to_i.to_s
+      File.open(temp_filename, "w") do |f|
+        f << ( @@vote_results << {:number => number, :vote => vote, :created_at => Time.now.to_i} ).to_yaml
+      end
+      File.copy(temp_filename, @@vote_results_file)
+      File.delete(temp_filename)
+    else
+        ( @@vote_results << {:number => number, :vote => vote, :created_at => Time.now.to_i} ).to_yaml
     end
-    File.copy(temp_filename, @@vote_results_file)
-    File.delete(temp_filename)
   end
 
   def self.already_voted?(number)
@@ -28,7 +32,12 @@ class VoteLogger
   private
 
   def self.load_previous_results
-    @@vote_results = YAML.load_file(@@vote_results_file)
+    if File.writeable? @@vote_results_file 
+      @@vote_results = YAML.load_file(@@vote_results_file)
+    else
+      @@dont_write = true
+      @@vote_results |= []
+    end
   end
 
 end
